PDF to LI-jpg/requirements.txt
PyMuPDF==1.18.19
Pillow==8.2.0
PyQt5==5.15.4
